refactor(cosmo_val): extract PSFSystematicsMixin (psf_systematics)
