+++ src/sp_validation/cosmo_val/psf_systematics.py
@@ -1,0 +1,649 @@
+# %%
+"""PSF systematics diagnostics.
+
+Rho/tau statistics, their fits, and PSF leakage tests (scale-dependent and
+object-wise) for catalogue validation.
+"""
+
+import os
+
+import matplotlib.pyplot as plt
+import numpy as np
+import treecorr
+from cs_util import plots as cs_plots
+from shear_psf_leakage import leakage
+from shear_psf_leakage import plots as psfleak_plots
+from shear_psf_leakage.rho_tau_stat import PSFErrorFit
+from uncertainties import ufloat
+
+from .rho_tau import (
+    get_rho_tau_w_cov,
+    get_samples,
+)
+
+
+class PSFSystematicsMixin:
+    def calculate_rho_tau_stats(self):
+        out_dir = f"{self.cc['paths']['output']}/rho_tau_stats"
+        if not os.path.exists(out_dir):
+            os.mkdir(out_dir)
+
+        self.print_start("Rho stats")
+        for ver in self.versions:
+            base = self.basename(ver)
+            rho_stat_handler, tau_stat_handler = get_rho_tau_w_cov(
+                self.cc,
+                ver,
+                self.treecorr_config,
+                out_dir,
+                base,
+                method=self.cov_estimate_method,
+                cov_rho=self.compute_cov_rho,
+                npatch=self.npatch,
+            )
+        self.print_done("Rho stats finished")
+
+        self._rho_stat_handler = rho_stat_handler
+        self._tau_stat_handler = tau_stat_handler
+
+    @property
+    def rho_stat_handler(self):
+        if not hasattr(self, "_rho_stat_handler"):
+            self.calculate_rho_tau_stats()
+        return self._rho_stat_handler
+
+    @property
+    def tau_stat_handler(self):
+        if not hasattr(self, "_tau_stat_handler"):
+            self.calculate_rho_tau_stats()
+        return self._tau_stat_handler
+
+    def plot_rho_stats(self, abs=False):
+        filenames = [f"rho_stats_{self.basename(ver)}.fits" for ver in self.versions]
+
+        savefig = "rho_stats.png"
+        self.rho_stat_handler.plot_rho_stats(
+            filenames,
+            self.colors,
+            self.versions,
+            savefig=savefig,
+            legend="outside",
+            abs=abs,
+            show=True,
+            close=True,
+        )
+
+        self.print_done(
+            "Rho stats plot saved to "
+            + f"{os.path.abspath(self.rho_stat_handler.catalogs._output)}/{savefig}",
+        )
+
+    def plot_tau_stats(self, plot_tau_m=False):
+        filenames = [f"tau_stats_{self.basename(ver)}.fits" for ver in self.versions]
+
+        savefig = "tau_stats.png"
+        self.tau_stat_handler.plot_tau_stats(
+            filenames,
+            self.colors,
+            self.versions,
+            savefig=savefig,
+            legend="outside",
+            plot_tau_m=plot_tau_m,
+            show=True,
+            close=True,
+        )
+
+        self.print_done(
+            "Tau stats plot saved to "
+            + f"{os.path.abspath(self.tau_stat_handler.catalogs._output)}/{savefig}",
+        )
+
+    def set_params_rho_tau(self, params, params_psf, survey="other"):
+        params = {**params}
+        if survey in ("DES", "SP_axel_v0.0", "SP_axel_v0.0_repr"):
+            params["patch_number"] = 120
+            print("DES, jackknife patch number = 120")
+        elif survey == "SP_axel_v0.0":
+            params["patch_number"] = 120
+            print("SP_Axel_v0.0, jackknife patch number =120")
+        elif survey == "SP_v1.4-P3" or survey == "SP_v1.4-P3_LFmask":
+            params["patch_number"] = 120
+            print("SP_v1.4, jackknife patch number =120")
+        else:
+            params["patch_number"] = 150
+
+        params["ra_PSF_col"] = params_psf["ra_col"]
+        params["dec_PSF_col"] = params_psf["dec_col"]
+        params["e1_PSF_col"] = params_psf["e1_PSF_col"]
+        params["e2_PSF_col"] = params_psf["e2_PSF_col"]
+        params["e1_star_col"] = params_psf["e1_star_col"]
+        params["e2_star_col"] = params_psf["e2_star_col"]
+        params["PSF_size"] = params_psf["PSF_size"]
+        params["star_size"] = params_psf["star_size"]
+        if survey != "DES":
+            params["PSF_flag"] = params_psf["PSF_flag"]
+            params["star_flag"] = params_psf["star_flag"]
+        params["ra_units"] = "deg"
+        params["dec_units"] = "deg"
+
+        params["w_col"] = self.cc[survey]["shear"]["w_col"]
+
+        return params
+
+    @property
+    def psf_fitter(self):
+        if not hasattr(self, "_psf_fitter"):
+            self._psf_fitter = PSFErrorFit(
+                self.rho_stat_handler,
+                self.tau_stat_handler,
+                self.rho_stat_handler.catalogs._output,
+            )
+        return self._psf_fitter
+
+    def calculate_rho_tau_fits(self):
+        assert self.rho_tau_method != "none"
+
+        # this initializes the rho_tau_fits attribute
+        self._rho_tau_fits = {"flat_sample_list": [], "result_list": [], "q_list": []}
+        quantiles = [1 - self.quantile, self.quantile]
+
+        self._xi_psf_sys = {}
+        for ver in self.versions:
+            params = self.set_params_rho_tau(
+                self.results[ver]._params,
+                self.cc[ver]["psf"],
+                survey=ver,
+            )
+
+            npatch = {"sim": 300, "jk": params["patch_number"]}.get(
+                self.cov_estimate_method, None
+            )
+
+            base = self.basename(ver)
+
+            flat_samples, result, q = get_samples(
+                self.psf_fitter,
+                ver,
+                base,
+                cov_type=self.cov_estimate_method,
+                apply_debias=npatch,
+                sampler=self.rho_tau_method,
+            )
+
+            self.rho_tau_fits["flat_sample_list"].append(flat_samples)
+            self.rho_tau_fits["result_list"].append(result)
+            self.rho_tau_fits["q_list"].append(q)
+
+            self.psf_fitter.load_rho_stat(f"rho_stats_{self.basename(ver)}.fits")
+            nbins = self.psf_fitter.rho_stat_handler._treecorr_config["nbins"]
+            xi_psf_sys_samples = np.array([]).reshape(0, nbins)
+
+            for i in range(len(flat_samples)):
+                xi_psf_sys = self.psf_fitter.compute_xi_psf_sys(flat_samples[i])
+                xi_psf_sys_samples = np.vstack([xi_psf_sys_samples, xi_psf_sys])
+
+            self._xi_psf_sys[ver] = {
+                "mean": np.mean(xi_psf_sys_samples, axis=0),
+                "var": np.var(xi_psf_sys_samples, axis=0),
+                "quantiles": np.quantile(xi_psf_sys_samples, quantiles, axis=0),
+            }
+
+    @property
+    def rho_tau_fits(self):
+        if not hasattr(self, "_rho_tau_fits"):
+            self.calculate_rho_tau_fits()
+        return self._rho_tau_fits
+
+    def plot_rho_tau_fits(self):
+        out_dir = self.rho_stat_handler.catalogs._output
+
+        savefig = f"{out_dir}/contours_tau_stat.png"
+        psfleak_plots.plot_contours(
+            self.rho_tau_fits["flat_sample_list"],
+            names=["x0", "x1", "x2"],
+            labels=[r"\alpha", r"\beta", r"\eta"],
+            savefig=savefig,
+            legend_labels=self.versions,
+            legend_loc="upper right",
+            contour_colors=self.colors,
+            markers={"x0": 0, "x1": 1, "x2": 1},
+            show=True,
+            close=True,
+        )
+        self.print_done(f"Tau contours plot saved to {os.path.abspath(savefig)}")
+
+        plt.figure(figsize=(15, 6))
+        for mcmc_result, ver, color, flat_sample in zip(
+            self.rho_tau_fits["result_list"],
+            self.versions,
+            self.colors,
+            self.rho_tau_fits["flat_sample_list"],
+        ):
+            self.psf_fitter.load_rho_stat(f"rho_stats_{self.basename(ver)}.fits")
+            for i in range(100):
+                self.psf_fitter.plot_xi_psf_sys(
+                    flat_sample[-i + 1], ver, color, alpha=0.1
+                )
+            self.psf_fitter.plot_xi_psf_sys(mcmc_result[1], ver, color)
+        plt.legend()
+        out_path = os.path.abspath(f"{out_dir}/xi_psf_sys_samples.png")
+        cs_plots.savefig(out_path, close_fig=False)
+        cs_plots.show()
+        self.print_done(f"xi_psf_sys samples plot saved to {out_path}")
+
+        plt.figure(figsize=(15, 6))
+        for mcmc_result, ver, color, flat_sample in zip(
+            self.rho_tau_fits["result_list"],
+            self.versions,
+            self.colors,
+            self.rho_tau_fits["flat_sample_list"],
+        ):
+            ls = self.cc[ver]["ls"]
+            theta = self.psf_fitter.rho_stat_handler.rho_stats["theta"]
+            xi_psf_sys = self.xi_psf_sys[ver]
+            plt.plot(theta, xi_psf_sys["mean"], linestyle=ls, color=color)
+            plt.plot(theta, xi_psf_sys["quantiles"][0], linestyle=ls, color=color)
+            plt.plot(theta, xi_psf_sys["quantiles"][1], linestyle=ls, color=color)
+            plt.fill_between(
+                theta,
+                xi_psf_sys["quantiles"][0],
+                xi_psf_sys["quantiles"][1],
+                color=color,
+                alpha=0.25,
+                label=ver,
+            )
+
+        plt.xscale("log")
+        plt.yscale("log")
+        plt.xlabel(r"$\theta$ [arcmin]")
+        plt.ylabel(r"$\xi^{\rm PSF}_{\rm sys}$")
+        plt.title(f"{1 - self.quantile:.1%}, {self.quantile:.1%} quantiles")
+        plt.legend()
+        out_path = os.path.abspath(f"{out_dir}/xi_psf_sys_quantiles.png")
+        cs_plots.savefig(out_path, close_fig=False)
+        cs_plots.show()
+        self.print_done(f"xi_psf_sys quantiles plot saved to {out_path}")
+
+        for mcmc_result, ver, flat_sample in zip(
+            self.rho_tau_fits["result_list"],
+            self.versions,
+            self.rho_tau_fits["flat_sample_list"],
+        ):
+            self.psf_fitter.load_rho_stat(f"rho_stats_{self.basename(ver)}.fits")
+            for yscale in ("linear", "log"):
+                out_path = os.path.abspath(
+                    f"{out_dir}/xi_psf_sys_terms_{yscale}_{ver}.png"
+                )
+                self.psf_fitter.plot_xi_psf_sys_terms(
+                    ver, mcmc_result[1], out_path, yscale=yscale, show=True
+                )
+                self.print_done(
+                    f"{yscale}-scale xi_psf_sys terms plot saved to {out_path}"
+                )
+
+    @property
+    def xi_psf_sys(self):
+        if not hasattr(self, "_xi_psf_sys"):
+            self.calculate_rho_tau_fits()
+        return self._xi_psf_sys
+
+    def set_params_leakage_scale(self, ver):
+        params_in = {}
+
+        # Set parameters
+        params_in["input_path_shear"] = self.cc[ver]["shear"]["path"]
+        params_in["input_path_PSF"] = self.cc[ver]["star"]["path"]
+        params_in["dndz_path"] = (
+            f"{self.cc['nz']['dndz']['path']}_{self.cc[ver]['pipeline']}_{self.cc['nz']['dndz']['blind']}.txt"
+        )
+        params_in["output_dir"] = f"{self.cc['paths']['output']}/leakage_{ver}"
+
+        # Note: for SP these are calibrated shear estimates
+        params_in["e1_col"] = self.cc[ver]["shear"]["e1_col"]
+        params_in["e2_col"] = self.cc[ver]["shear"]["e2_col"]
+        params_in["w_col"] = self.cc[ver]["shear"]["w_col"]
+        params_in["R11"] = None if ver != "DES" else self.cc[ver]["shear"]["R11"]
+        params_in["R22"] = None if ver != "DES" else self.cc[ver]["shear"]["R22"]
+
+        params_in["ra_star_col"] = self.cc[ver]["star"]["ra_col"]
+        params_in["dec_star_col"] = self.cc[ver]["star"]["dec_col"]
+        params_in["e1_PSF_star_col"] = self.cc[ver]["star"]["e1_col"]
+        params_in["e2_PSF_star_col"] = self.cc[ver]["star"]["e2_col"]
+
+        params_in["theta_min_amin"] = self.theta_min
+        params_in["theta_max_amin"] = self.theta_max
+        params_in["n_theta"] = self.nbins
+
+        params_in["verbose"] = False
+
+        return params_in
+
+    def set_params_leakage_object(self, ver):
+        params_in = {}
+
+        # Set parameters
+        params_in["input_path_shear"] = self.cc[ver]["shear"]["path"]
+        params_in["output_dir"] = f"{self.cc['paths']['output']}/leakage_{ver}"
+
+        # Note: for SP these are calibrated shear estimates
+        params_in["e1_col"] = self.cc[ver]["shear"]["e1_col"]
+        params_in["e2_col"] = self.cc[ver]["shear"]["e2_col"]
+        params_in["w_col"] = self.cc[ver]["shear"]["w_col"]
+
+        if (
+            "e1_PSF_col" in self.cc[ver]["shear"]
+            and "e2_PSF_col" in self.cc[ver]["shear"]
+        ):
+            params_in["e1_PSF_col"] = self.cc[ver]["shear"]["e1_PSF_col"]
+            params_in["e2_PSF_col"] = self.cc[ver]["shear"]["e2_PSF_col"]
+        else:
+            raise KeyError(
+                "Keys 'e1_PSF_col' and 'e2_PSF_col' not found in"
+                + f" shear yaml entry for version {ver}"
+            )
+
+        params_in["verbose"] = False
+
+        return params_in
+
+    def calculate_scale_dependent_leakage(self):
+        self.print_start("Calculating scale-dependent leakage:")
+        for ver in self.versions:
+            self.print_magenta(ver)
+            results = self.results[ver]
+
+            output_base_path = os.path.abspath(
+                f"{self.cc['paths']['output']}/leakage_{ver}/xi_for_leak_scale"
+            )
+            output_path_ab = f"{output_base_path}_a_b.txt"
+            output_path_aa = f"{output_base_path}_a_a.txt"
+            with self.results[ver].temporarily_read_data():
+                if os.path.exists(output_path_ab) and os.path.exists(output_path_aa):
+                    self.print_green(
+                        f"Skipping computation, reading {output_path_ab} and "
+                        f"{output_path_aa} instead"
+                    )
+
+                    results.r_corr_gp = treecorr.GGCorrelation(self.treecorr_config)
+                    results.r_corr_gp.read(output_path_ab)
+
+                    results.r_corr_pp = treecorr.GGCorrelation(self.treecorr_config)
+                    results.r_corr_pp.read(output_path_aa)
+
+                else:
+                    results.compute_corr_gp_pp_alpha(output_base_path=output_base_path)
+
+                results.do_alpha(fast=True)
+                results.do_xi_sys()
+
+        self.print_done("Finished scale-dependent leakage calculation.")
+
+    def plot_scale_dependent_leakage(self):
+        if not hasattr(self.results[self.versions[0]], "r_corr_gp"):
+            self.calculate_scale_dependent_leakage()
+
+        theta = []
+        y = []
+        yerr = []
+        labels = []
+        colors = []
+        linestyles = []
+        markers = []
+
+        for ver in self.versions:
+            if hasattr(self.results[ver], "r_corr_gp"):
+                theta.append(self.results[ver].r_corr_gp.meanr)
+                y.append(self.results[ver].alpha_leak)
+                yerr.append(self.results[ver].sig_alpha_leak)
+                labels.append(ver)
+                colors.append(self.cc[ver]["colour"])
+                linestyles.append(self.cc[ver]["ls"])
+                markers.append(self.cc[ver]["marker"])
+
+        if len(theta) > 0:
+            # Log x
+            out_path = os.path.abspath(
+                f"{self.cc['paths']['output']}/alpha_leak_log.png"
+            )
+
+            title = r"$\alpha$ leakage"
+            xlabel = r"$\theta$ [arcmin]"
+            ylabel = r"$\alpha(\theta)$"
+            cs_plots.plot_data_1d(
+                theta,
+                y,
+                yerr,
+                title,
+                xlabel,
+                ylabel,
+                out_path=None,
+                xlog=True,
+                xlim=[self.theta_min_plot, self.theta_max_plot],
+                ylim=self.ylim_alpha,
+                labels=labels,
+                colors=colors,
+                linestyles=linestyles,
+                shift_x=True,
+            )
+            cs_plots.savefig(out_path, close_fig=False)
+            cs_plots.show()
+            self.print_done(f"Log-scale alpha leakage plot saved to {out_path}")
+
+            # Lin x
+            out_path = os.path.abspath(
+                f"{self.cc['paths']['output']}/alpha_leak_lin.png"
+            )
+
+            title = r"$\alpha$ leakage"
+            xlabel = r"$\theta$ [arcmin]"
+            ylabel = r"$\alpha(\theta)$"
+            cs_plots.plot_data_1d(
+                theta,
+                y,
+                yerr,
+                title,
+                xlabel,
+                ylabel,
+                out_path=None,
+                xlog=False,
+                xlim=[-10, self.theta_max_plot],
+                ylim=self.ylim_alpha,
+                labels=labels,
+                colors=colors,
+                linestyles=linestyles,
+                shift_x=False,
+            )
+            cs_plots.savefig(out_path, close_fig=False)
+            cs_plots.show()
+            self.print_done(f"Lin-scale alpha leakage plot saved to {out_path}")
+
+        # Plot xi_sys
+        y = []
+        yerr = []
+        colors = []
+        linestyles = []
+
+        for ver in self.versions:
+            if hasattr(self.results[ver], "C_sys_p"):
+                y.append(self.results[ver].C_sys_p)
+                yerr.append(self.results[ver].C_sys_std_p)
+                labels.append(ver)
+                colors.append(self.cc[ver]["colour"])
+                linestyles.append(self.cc[ver]["ls"])
+
+        if len(y) > 0:
+            xlabel = r"$\theta$ [arcmin]"
+            ylabel = r"$\xi^{\rm sys}_+(\theta)$"
+            title = "Cross-correlation leakage"
+            out_path = os.path.abspath(f"{self.cc['paths']['output']}/xi_sys_p.png")
+            cs_plots.plot_data_1d(
+                theta,
+                y,
+                yerr,
+                title,
+                xlabel,
+                ylabel,
+                out_path=None,
+                labels=labels,
+                xlog=True,
+                xlim=[self.theta_min_plot, self.theta_max_plot],
+                colors=colors,
+                linestyles=linestyles,
+                # shift_x=True,
+            )
+            cs_plots.savefig(out_path, close_fig=False)
+            cs_plots.show()
+            self.print_done(f"xi_sys_plus plot saved to {out_path}")
+
+        y = []
+        yerr = []
+        for ver in self.versions:
+            if hasattr(self.results[ver], "C_sys_m"):
+                y.append(self.results[ver].C_sys_m)
+                yerr.append(self.results[ver].C_sys_std_m)
+
+        if len(y) > 0:
+            xlabel = r"$\theta$ [arcmin]"
+            ylabel = r"$\xi^{\rm sys}_-(\theta)$"
+            title = "Cross-correlation leakage"
+            out_path = os.path.abspath(f"{self.cc['paths']['output']}/xi_sys_m.png")
+            cs_plots.plot_data_1d(
+                theta,
+                y,
+                yerr,
+                title,
+                xlabel,
+                ylabel,
+                out_path=None,
+                labels=labels,
+                xlog=True,
+                xlim=[self.theta_min_plot, self.theta_max_plot],
+                ylim=[-1e-7, 1e-6],
+                colors=colors,
+                linestyles=linestyles,
+                # shift_x=True,
+            )
+            cs_plots.savefig(out_path, close_fig=False)
+            cs_plots.show()
+            self.print_done(f"xi_sys_minus plot saved to {out_path}")
+
+    def calculate_objectwise_leakage(self):
+        if not hasattr(self.results[self.versions[0]], "alpha_leak_mean"):
+            self.calculate_scale_dependent_leakage()
+
+        self.print_start("Object-wise leakage:")
+        mix = True
+        order = "lin"
+        for ver in self.versions:
+            self.print_magenta(ver)
+
+            results_obj = self.results_objectwise[ver]
+            results_obj.check_params()
+            results_obj.update_params()
+            results_obj.prepare_output()
+
+            # Skip read_data() and copy catalogue from scale leakage instance instead
+            # results_obj._dat = self.results[ver].dat_shear
+
+            out_base = results_obj.get_out_base(mix, order)
+            out_path = f"{out_base}.pkl"
+            if os.path.exists(out_path):
+                self.print_green(
+                    f"Skipping object-wise leakage, file {out_path} exists"
+                )
+                results_obj.par_best_fit = leakage.read_from_file(out_path)
+            else:
+                self.print_cyan("Computing object-wise leakage regression")
+
+            # Run
+            with results_obj.temporarily_read_data():
+                try:
+                    results_obj.PSF_leakage()
+                except KeyError as e:
+                    print(f"{e}\nExpected key is missing from catalog.")
+                    # remove the results object for this version
+                    self.results_objectwise.pop(ver)
+
+        # Gather coefficients
+        leakage_coeff = {}
+        for ver in self.results_objectwise:
+            leakage_coeff[ver] = {}
+            results = self.results[ver]
+            results_obj = self.results_objectwise[ver]
+            # Object-wise leakage
+            leakage_coeff[ver]["a11"] = ufloat(
+                results_obj.par_best_fit["a11"].value,
+                results_obj.par_best_fit["a11"].stderr,
+            )
+            leakage_coeff[ver]["a22"] = ufloat(
+                results_obj.par_best_fit["a22"].value,
+                results_obj.par_best_fit["a22"].stderr,
+            )
+            leakage_coeff[ver]["aii_mean"] = 0.5 * (
+                leakage_coeff[ver]["a11"] + leakage_coeff[ver]["a22"]
+            )
+
+            # Scale-dependent leakage: mean
+            leakage_coeff[ver]["alpha_mean"] = ufloat(
+                results.alpha_leak_mean, results.alpha_leak_std
+            )
+            # Scale-dependent leakage: value at smallest scale
+            leakage_coeff[ver]["alpha_1"] = ufloat(
+                results.alpha_leak[0], results.sig_alpha_leak[0]
+            )
+            # Scale-dependent leakage: value extrapolated to 0 using affine model
+            leakage_coeff[ver]["alpha_0"] = ufloat(
+                results.alpha_affine_best_fit["c"].value,
+                results.alpha_affine_best_fit["c"].stderr,
+            )
+
+        self.leakage_coeff = leakage_coeff
+
+    def plot_objectwise_leakage(self):
+        if not hasattr(self, "leakage_coeff"):
+            self.calculate_objectwise_leakage()
+
+        self.print_start("Plotting object-wise leakage:")
+        cs_plots.figure(figsize=(15, 15))
+
+        linestyles = ["-", "--", ":"]
+        fillstyles = ["full", "none", "left", "right", "bottom", "top"]
+
+        for ver in self.results_objectwise:
+            label = ver
+            for key, ls, fs in zip(
+                ["alpha_mean", "alpha_1", "alpha_0"], linestyles, fillstyles
+            ):
+                x = self.leakage_coeff[ver]["aii_mean"].nominal_value
+                dx = self.leakage_coeff[ver]["aii_mean"].std_dev
+                y = self.leakage_coeff[ver][key].nominal_value
+                dy = self.leakage_coeff[ver][key].std_dev
+
+                eb = plt.errorbar(
+                    x,
+                    y,
+                    xerr=dx,
+                    yerr=dy,
+                    fmt=self.cc[ver]["marker"],
+                    color=self.cc[ver]["colour"],
+                    fillstyle=fs,
+                    label=label,
+                )
+                label = None
+                eb[-1][0].set_linestyle(ls)
+
+        # y=x line
+        xlim = 0.02
+        x = [-xlim, xlim]
+        y = x
+        plt.plot(x, y, "k:", linewidth=0.5)
+
+        plt.legend()
+        plt.xlabel(r"tr $a$ (object-wise)")
+        plt.ylabel(r"$\alpha$ (scale-dependent)")
+        out_path = os.path.abspath(
+            f"{self.cc['paths']['output']}/leakage_coefficients.png"
+        )
+        cs_plots.savefig(out_path, close_fig=False)
+        cs_plots.show()
+        self.print_done(f"Object-wise leakage coefficients plot saved to {out_path}")
